Updated blank & comet pattern

+++ src/patterns/CrazyPattern.cpp
@@ -4,17 +4,29 @@
 namespace Loml {
     void CrazyPattern::DisplayImpl(LEDStrip& led) {
         constexpr static auto colorWheel = [](uint8_t pos) {
-            constexpr static auto max = 255;
+            constexpr static auto max = static_cast<uint8_t>(255);
             pos = max - pos;
 
             if (pos < 85) {
-                return RgbColor{max - pos * 3, 0, pos * 3};
+                return RgbColor{
+                    static_cast<uint8_t>(max - pos * 3), 
+                    static_cast<uint8_t>(0),
+                    static_cast<uint8_t>(pos * 3)
+                };
             } else if (pos < 170) {
                 pos -= 85;
-                return RgbColor{0, pos * 3, max - pos * 3};
+                return RgbColor{
+                    static_cast<uint8_t>(0), 
+                    static_cast<uint8_t>(pos * 3), 
+                    static_cast<uint8_t>(max - pos * 3)
+                };
             } else {
                 pos -= 170;
-                return RgbColor{pos * 3, max - pos * 3, 0};
+                return RgbColor{
+                    static_cast<uint8_t>(pos * 3), 
+                    static_cast<uint8_t>(max - pos * 3), 
+                    static_cast<uint8_t>(0)
+                };
             }
         };
